backend/pyproject.toml
[project]
name = "hack4bengal2025"
version = "0.1.0"
description = "Add your description here"
authors = [{ name="Shubhajit Halder", email="shubhajith85@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.2",
    "bcrypt>=4.3.0",
    "celery>=5.5.1",
    "fastapi[standard]>=0.115.12",
    "groq>=0.22.0",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-groq>=0.3.2",
    "pydantic-settings>=2.9.0",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "pywebpush>=2.0.3",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.40",
    "tavily-python>=0.5.4",
    "langchain-chroma>=0.2.3",
    "pypdf>=5.4.0",
    "asyncpg>=0.30.0",
    "psycopg2>=2.9.10",
    "pytz>=2025.2",
    "passlib>=1.7.4",
    "langchain-ollama>=0.3.2",
    "scikit-learn>=1.6.1",
    "agno>=1.3.4",
    "lancedb>=0.21.2",
    "pandas>=2.2.3",
    "llama-index>=0.12.31",
    "llama-index-llms-groq>=0.3.1",
    "llama-index-llms-ollama>=0.5.4",
    "llama-index-embeddings-ollama>=0.6.0",
    "llama-index-vector-stores-chroma>=0.4.1",
    "crewai>=0.114.0",
    "crewai-tools>=0.0.1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.11.6",
]
